\name{DATASET.1}
\alias{DATASET.1}
\docType{data}
\title{
A 50 x 5 data set
}
\description{
One of two 50 x 5 data sets as mentioned on p. 135 of Jouan-Rimbaud et al (1998).
See DATASET.2 for the other 50 x 5 data set.
}
\usage{data(DATASET.1)}
\format{
  The format is:
 num [1:50, 1:5] 19.851 7.526 2.123 0.945 0.726 ...
}
\details{
Variable 1: 50 values uniformly distributed between 0 and 20 + noise.
Variable 2: 50 values uniformly distributed between 2 and 20.
Variable 3: 50 values uniformly distributed between 10 and 20.
Variable 4: 50 values uniformly distributed between 5 and 20.
Variable 5: 50 values uniformly distributed between 6 and 20.
}
\source{
Page 135 of: Jouan-Rimbaud D, Massart DL, Saby CA, Puel C: Determination of the representativity between two multidimensional data sets by a comparison of their structure. Chemometrics and Intelligent Laboratory Systems 40 (1998) 129-144.
}
\examples{
data(DATASET.1)
}
%\keyword{datasets}
