\name{DATASET.4}
\alias{DATASET.4}
\docType{data}
\title{
A 50 x 10 data set
}
\description{
One of two 50 x 10 data sets as mentioned on p. 135 of Jouan-Rimbaud et al (1998).
See DATASET.3 for the other 50 x 10 data set.
}
\usage{data(DATASET.4)}
\format{
  The format is:
 num [1:50, 1:10] 2.72 12.05 6.5 12.27 16.03 ...
}
\details{
Variable 1: 50 values uniformly distributed between 0 and 20 + noise.
Variable 2: 50 values uniformly distributed between 2 and 20.
Variable 3: 50 values uniformly distributed between 10 and 20.
Variable 4: 50 values uniformly distributed between 5 and 20.
Variable 5: 50 values uniformly distributed between 6 and 20.
Variables 6-10: five variables with values uniformly distributed between -0.5 and +0.5
}
\source{
Page 135 of: Jouan-Rimbaud D, Massart DL, Saby CA, Puel C: Determination of the representativity between two multidimensional data sets by a comparison of their structure. Chemometrics and Intelligent Laboratory Systems 40 (1998) 129-144.
}
\examples{
data(DATASET.4)
}
%\keyword{datasets}
