\name{represent-package}
\alias{represent-package}
\alias{represent}
\docType{package}
\title{
Determine the representativity of two multidimensional data sets
}
\description{
This package contains the workhorse function jrparams(), as well as two helper functions Mboxtest() and JRsMahaldist(), and four example data sets.
The jrparams() function computes the values of three types of parameters that assess the representativity of two multidimensional data sets.
These parameters and the example data sets are described in a publication by Jouan-Rimbaud et al (1998).
}
\details{
  \tabular{ll}{
  Package: \tab represent\cr
  Type: \tab Package\cr
  Version: \tab 1.0.1\cr
  Date: \tab 2023-11-02\cr
  License: \tab GPL-3\cr#
  }
}
\author{
Harmen Draisma

Maintainer: Harmen Draisma <h.h.m.draisma@gmail.com>
}
\references{
Jouan-Rimbaud D, Massart DL, Saby CA, Puel C: Determination of the representativity between two multidimensional data sets by a comparison of their structure. Chemometrics and Intelligent Laboratory Systems 40 (1998) 129-144.
}
%\keyword{ package }
